requirements.txt
cycler==0.11.0
fonttools==4.37.1
imageio==2.21.3
kiwisolver==1.4.4
matplotlib==3.5.3
networkx==2.8.6
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
pytesseract==0.3.10
python-dateutil==2.8.2
PyWavelets==1.3.0
scikit-image==0.19.3
scipy==1.9.1
six==1.16.0
tifffile==2022.8.12
